Remove obsolete example pages from docs.
