+++ Docs/quantlib.doxy
@@ -110,16 +110,12 @@ FILE_PATTERNS          = *.docs \
                          *.h
 RECURSIVE              = YES
 EXCLUDE                = ../ql/config.hpp \
-                         ../ql/config.bcc.hpp \
-                         ../ql/config.decc.hpp \
-                         ../ql/config.msvc.hpp \
-                         ../ql/config.mwcw.hpp
+                         ../ql/config.msvc.hpp
 EXCLUDE_SYMLINKS       = NO
 EXCLUDE_PATTERNS       = */all.hpp
 EXCLUDE_SYMBOLS        =
-EXAMPLE_PATH           = Examples \
-                         ../Examples
-EXAMPLE_PATTERNS       = *.cpp
+EXAMPLE_PATH           = 
+EXAMPLE_PATTERNS       = 
 EXAMPLE_RECURSIVE      = YES
 IMAGE_PATH             =
 INPUT_FILTER           =
